[refactor] move user APIs to JavaSupport
while internal ones (some of which are public but deprecated) stay at
the Impl sub-class
