+++ core/src/main/java/org/jruby/javasupport/JavaSupportImpl.java
@@ -34,7 +34,6 @@
 
 package org.jruby.javasupport;
 
-import java.lang.reflect.Member;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -42,246 +41,29 @@ import java.util.Map;
 import java.util.Set;
 
 import org.jruby.Ruby;
-import org.jruby.RubyClass;
-import org.jruby.RubyModule;
-import org.jruby.exceptions.RaiseException;
-import org.jruby.exceptions.Unrescuable;
-import org.jruby.runtime.Helpers;
 import org.jruby.util.ArraySupport;
+import org.jruby.util.WeakIdentityHashMap;
 import org.jruby.util.collections.ClassValue;
 import org.jruby.javasupport.binding.AssignedName;
 import org.jruby.javasupport.proxy.JavaProxyClass;
-import org.jruby.javasupport.util.ObjectProxyCache;
-import org.jruby.runtime.builtin.IRubyObject;
-import org.jruby.util.WeakIdentityHashMap;
 
-import static org.jruby.javasupport.Java.initCause;
-
+/**
+ * Internal APIs of {@link JavaSupport}.
+ */
 public class JavaSupportImpl extends JavaSupport {
 
-    private final ObjectProxyCache<IRubyObject,RubyClass> objectProxyCache =
-            // TODO: specifying soft refs, may want to compare memory consumption,
-            // behavior with weak refs (specify WEAK in place of SOFT below)
-            new ObjectProxyCache<IRubyObject,RubyClass>(ObjectProxyCache.ReferenceType.WEAK) {
-
-                public IRubyObject allocateProxy(Object javaObject, RubyClass clazz) {
-                    return Java.allocateProxy(javaObject, clazz);
-                }
-            };
-
-    private final ClassValue<Map<String, AssignedName>> staticAssignedNames;
-    private final ClassValue<Map<String, AssignedName>> instanceAssignedNames;
-
-    private RubyModule javaModule;
-    private RubyModule javaUtilitiesModule;
-    private RubyModule javaArrayUtilitiesModule;
-    private RubyClass javaObjectClass;
-    private JavaClass objectJavaClass;
-    private RubyClass javaClassClass;
-    private RubyClass javaPackageClass;
-    private RubyClass javaArrayClass;
-    private RubyClass javaProxyClass;
-    private RubyClass arrayJavaProxyCreatorClass;
-    private RubyClass javaFieldClass;
-    private RubyClass javaMethodClass;
-    private RubyClass javaConstructorClass;
-    private RubyModule javaInterfaceTemplate;
-    private RubyClass arrayProxyClass;
-    private RubyClass concreteProxyClass;
-    private RubyClass mapJavaProxy;
-    private RubyClass javaProxyConstructorClass;
+    private final ClassValue<Map<String, AssignedName>> staticAssignedNames =
+            ClassValue.newInstance(klass -> new HashMap<>(8, 1));
+    private final ClassValue<Map<String, AssignedName>> instanceAssignedNames =
+            ClassValue.newInstance(klass -> new HashMap<>(8, 1));
 
     public JavaSupportImpl(final Ruby runtime) {
         super(runtime);
-
-        this.staticAssignedNames = ClassValue.newInstance(klass -> new HashMap<>(8, 1));
-        this.instanceAssignedNames = ClassValue.newInstance(klass -> new HashMap<>(8, 1));
     }
 
-    @Override
     @Deprecated
-    public Class loadJavaClassVerbose(String className) {
-        try {
-            return loadJavaClass(className);
-        } catch (ClassNotFoundException ex) {
-            throw initCause(runtime.newNameError("cannot load Java class " + className, className, ex), ex);
-        } catch (ExceptionInInitializerError ex) {
-            throw initCause(runtime.newNameError("cannot initialize Java class " + className, className, ex), ex);
-        } catch (LinkageError ex) {
-            throw initCause(runtime.newNameError("cannot link Java class " + className + ", probable missing dependency: " + ex.getLocalizedMessage(), className, ex), ex);
-        } catch (SecurityException ex) {
-            if (runtime.isVerbose()) ex.printStackTrace(runtime.getErrorStream());
-            throw initCause(runtime.newSecurityError(ex.getLocalizedMessage()), ex);
-        }
-    }
-
-    @Override
-    @Deprecated
-    public Class loadJavaClassQuiet(String className) {
-        try {
-            return loadJavaClass(className);
-        } catch (ClassNotFoundException ex) {
-            throw initCause(runtime.newNameError("cannot load Java class " + className, className, ex, false), ex);
-        } catch (ExceptionInInitializerError ex) {
-            throw initCause(runtime.newNameError("cannot initialize Java class " + className, className, ex, false), ex);
-        } catch (LinkageError ex) {
-            throw initCause(runtime.newNameError("cannot link Java class " + className, className, ex, false), ex);
-        } catch (SecurityException ex) {
-            throw initCause(runtime.newSecurityError(ex.getLocalizedMessage()), ex);
-        }
-    }
-
-    public void handleNativeException(Throwable exception, Member target) {
-        if ( exception instanceof RaiseException ) {
-            // allow RaiseExceptions to propagate
-            throw (RaiseException) exception;
-        }
-        if (exception instanceof Unrescuable) {
-            // allow "unrescuable" flow-control exceptions to propagate
-            if ( exception instanceof Error ) {
-                throw (Error) exception;
-            }
-            if ( exception instanceof RuntimeException ) {
-                throw (RuntimeException) exception;
-            }
-        }
-        // rethrow original
-        Helpers.throwException(exception);
-    }
-
-    public ObjectProxyCache<IRubyObject,RubyClass> getObjectProxyCache() {
-        return objectProxyCache;
-    }
-
-    // not synchronizing these methods, no harm if these values get set more
-    // than once.
-    // (also note that there's no chance of getting a partially initialized
-    // class/module, as happens-before is guaranteed by volatile write/read
-    // of constants table.)
-
     public Map<String, JavaClass> getNameClassMap() {
         return Collections.emptyMap();
-    }
-
-    public RubyModule getJavaModule() {
-        RubyModule module;
-        if ((module = javaModule) != null) return module;
-        return javaModule = runtime.getModule("Java");
-    }
-
-    public RubyModule getJavaUtilitiesModule() {
-        RubyModule module;
-        if ((module = javaUtilitiesModule) != null) return module;
-        return javaUtilitiesModule = runtime.getModule("JavaUtilities");
-    }
-
-    public RubyModule getJavaArrayUtilitiesModule() {
-        RubyModule module;
-        if ((module = javaArrayUtilitiesModule) != null) return module;
-        return javaArrayUtilitiesModule = runtime.getModule("JavaArrayUtilities");
-    }
-
-    public RubyClass getJavaObjectClass() {
-        RubyClass clazz;
-        if ((clazz = javaObjectClass) != null) return clazz;
-        return javaObjectClass = getJavaModule().getClass("JavaObject");
-    }
-
-    public RubyClass getJavaProxyConstructorClass() {
-        RubyClass clazz;
-        if ((clazz = javaProxyConstructorClass) != null) return clazz;
-        return javaProxyConstructorClass = getJavaModule().getClass("JavaProxyConstructor");
-    }
-
-    @Deprecated // no longer used
-    public JavaClass getObjectJavaClass() {
-        JavaClass clazz;
-        if ((clazz = objectJavaClass) != null) return clazz;
-        return objectJavaClass = JavaClass.get(runtime, Object.class);
-    }
-
-    @Deprecated
-    public void setObjectJavaClass(JavaClass objectJavaClass) {
-        // noop
-    }
-
-    public RubyClass getJavaArrayClass() {
-        RubyClass clazz;
-        if ((clazz = javaArrayClass) != null) return clazz;
-        return javaArrayClass = getJavaModule().getClass("JavaArray");
-    }
-
-    public RubyClass getJavaClassClass() {
-        RubyClass clazz;
-        if ((clazz = javaClassClass) != null) return clazz;
-        return javaClassClass = getJavaModule().getClass("JavaClass");
-    }
-
-    public RubyClass getJavaPackageClass() {
-        return javaPackageClass;
-    }
-
-    public void setJavaPackageClass(RubyClass javaPackageClass) {
-        this.javaPackageClass = javaPackageClass;
-    }
-
-    public RubyModule getJavaInterfaceTemplate() {
-        RubyModule module;
-        if ((module = javaInterfaceTemplate) != null) return module;
-        return javaInterfaceTemplate = runtime.getModule("JavaInterfaceTemplate");
-    }
-
-    @Deprecated
-    public RubyModule getPackageModuleTemplate() {
-        return null; // no longer used + has been deprecated since ~ 9.1
-    }
-
-    public RubyClass getJavaProxyClass() {
-        RubyClass clazz;
-        if ((clazz = javaProxyClass) != null) return clazz;
-        return javaProxyClass = runtime.getClass("JavaProxy");
-    }
-
-    public RubyClass getArrayJavaProxyCreatorClass() {
-        RubyClass clazz;
-        if ((clazz = arrayJavaProxyCreatorClass) != null) return clazz;
-        return arrayJavaProxyCreatorClass = runtime.getClass("ArrayJavaProxyCreator");
-    }
-
-    public RubyClass getConcreteProxyClass() {
-        RubyClass clazz;
-        if ((clazz = concreteProxyClass) != null) return clazz;
-        return concreteProxyClass = runtime.getClass("ConcreteJavaProxy");
-    }
-
-    public RubyClass getMapJavaProxyClass() {
-        RubyClass clazz;
-        if ((clazz = mapJavaProxy) != null) return clazz;
-        return mapJavaProxy = runtime.getClass("MapJavaProxy");
-    }
-
-    public RubyClass getArrayProxyClass() {
-        RubyClass clazz;
-        if ((clazz = arrayProxyClass) != null) return clazz;
-        return arrayProxyClass = runtime.getClass("ArrayJavaProxy");
-    }
-
-    public RubyClass getJavaFieldClass() {
-        RubyClass clazz;
-        if ((clazz = javaFieldClass) != null) return clazz;
-        return javaFieldClass = getJavaModule().getClass("JavaField");
-    }
-
-    public RubyClass getJavaMethodClass() {
-        RubyClass clazz;
-        if ((clazz = javaMethodClass) != null) return clazz;
-        return javaMethodClass = getJavaModule().getClass("JavaMethod");
-    }
-
-    public RubyClass getJavaConstructorClass() {
-        RubyClass clazz;
-        if ((clazz = javaConstructorClass) != null) return clazz;
-        return javaConstructorClass = getJavaModule().getClass("JavaConstructor");
     }
 
     public ClassValue<Map<String, AssignedName>> getStaticAssignedNames() {
@@ -316,15 +98,19 @@ public class JavaSupportImpl extends JavaSupport {
     // cache of all JavaProxyClass objects created for this runtime
     private final Map<ProxyClassKey, JavaProxyClass> javaProxyClasses = new HashMap<>();
 
-    @Override
-    final protected JavaProxyClass fetchJavaProxyClass(ProxyClassKey classKey) {
+    /**
+     * a replacement for {@link #getJavaProxyClassCache()} API
+     */
+    protected JavaProxyClass fetchJavaProxyClass(ProxyClassKey classKey) {
         synchronized (javaProxyClasses) {
             return javaProxyClasses.get(classKey);
         }
     }
 
-    @Override
-    final protected JavaProxyClass saveJavaProxyClass(ProxyClassKey classKey, JavaProxyClass klass) {
+    /**
+     * a replacement for {@link #getJavaProxyClassCache()} API
+     */
+    protected JavaProxyClass saveJavaProxyClass(ProxyClassKey classKey, JavaProxyClass klass) {
         synchronized (javaProxyClasses) {
             JavaProxyClass existing = javaProxyClasses.get(classKey);
             if ( existing != null ) return existing;
@@ -335,12 +121,71 @@ public class JavaSupportImpl extends JavaSupport {
 
     // internal helper to access non-public fetch method
     public static JavaProxyClass fetchJavaProxyClass(final Ruby runtime, ProxyClassKey classKey) {
-        return runtime.getJavaSupport().fetchJavaProxyClass(classKey);
+        return ((JavaSupportImpl) runtime.getJavaSupport()).fetchJavaProxyClass(classKey);
     }
 
     // internal helper to access non-public save method
     public static JavaProxyClass saveJavaProxyClass(final Ruby runtime, ProxyClassKey classKey, JavaProxyClass klass) {
-        return runtime.getJavaSupport().saveJavaProxyClass(classKey, klass);
+        return ((JavaSupportImpl) runtime.getJavaSupport()).saveJavaProxyClass(classKey, klass);
+    }
+
+    /**
+     * @note Internal API - subject to change!
+     */
+    public static final class ProxyClassKey {
+        final Class superClass;
+        final Class[] interfaces;
+        final Set<String> names; // "usable" method names - assumed immutable
+
+        private ProxyClassKey(Class superClass, Class[] interfaces, Set<String> names) {
+            this.superClass = superClass;
+            this.interfaces = interfaces;
+            this.names = names;
+        }
+
+        public static ProxyClassKey getInstance(Class superClass, Class[] interfaces, Set<String> names) {
+            return new ProxyClassKey(superClass, interfaces, names);
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if ( obj instanceof ProxyClassKey ) {
+                final ProxyClassKey that = (ProxyClassKey) obj;
+                if (this.superClass != that.superClass) return false;
+
+                if (this.names.size() != that.names.size()) return false;
+                if ( ! this.names.equals(that.names) ) return false;
+
+                final int len = this.interfaces.length;
+                if (len != that.interfaces.length) return false;
+                // order is not important :
+                for ( int i = 0; i < len; i++ ) {
+                    final Class iface = this.interfaces[i];
+                    boolean ifaceFound = false;
+                    for ( int j = 0; j < len; j++ ) {
+                        if ( iface == that.interfaces[j] ) {
+                            ifaceFound = true; break;
+                        }
+                    }
+                    if ( ! ifaceFound ) return false;
+                }
+                return true;
+            }
+            return false;
+        }
+
+        private int hash;
+
+        @Override
+        public int hashCode() {
+            int hash = this.hash;
+            if (hash != 0) return hash;
+
+            for ( int i = 0; i < interfaces.length; i++ ) {
+                hash += interfaces[i].hashCode();
+            }
+            return this.hash = (hash * superClass.hashCode()) ^ this.names.hashCode();
+        }
     }
 
     @Deprecated
